Add shared folder in PSC service
This pattern should be repeated in every service. There are some types (the schema for entities stored in the database) that are needed by the webservice, the bulk loader, and the stream updater. Uses typescript project references to manage. Will need Dockerfiles to be moved to the top level to facilitate copying the shared directory.

Signed-off-by: Brian Evans <ebrian101@gmail.com>