Change officer storage PK to include role and appointment date
Because it was discovered that a single person can be appointed to the same company for different roles simultaneously and a person can be appointed for a role and then resign, and then be appointed again, so the primary key needs to include appointment date and officer role.

At the same time any outstanding camelCase keys were changed to snake_case.

Signed-off-by: Brian Evans <ebrian101@gmail.com>