Company updater now uses bulk operations on a timer
Instead of individual inserts, it groups events into bulk write operations in Mongo.
A bulk write is performed every 10 seconds, and whenever the buffer of 1000 events fills up.
This should alleviate load when reconnecting to the stream after being disconnected for a while
and there are tens or hundreds of thousands of events queued up.

Signed-off-by: Brian Evans <ebrian101@gmail.com>